Store rasters in half precision
A raster is 0/1 planes and one ramp. fp16 is exact for the first and
keeps about three decimals of the second, which is above where an
ablation could still tell the difference: continuous fp32 reached policy
CE 2.8211 against 2.8441 for a 256-level quantization, so the margin sits
between fp16 and uint8 rather than below fp16.

Halving the footprint is what lets the replay window hold sixteen shards
rather than five. With the compact raster that is 15 GB against a machine
with 46 free, where ten channels in fp32 needed 75 GB for the same window.

The staging path already mirrors the source dtype, so the host buffers and
the device copy both halve; only the model boundary widens.

One test asserted five decimal places on a channel whose value is 0.2,
which fp16 cannot represent exactly. The observed error is one fp16 ulp,
so the bound moves to four places rather than the assertion being dropped.

+++ training/vgo_training/dataset.py
@@ -649,7 +649,13 @@ def load_dataset(path: str | Path) -> RasterDataset:
         raise ValueError("selected replay action is absent from the policy mask")
 
     return RasterDataset(
-        states=torch.from_numpy(states),
+        # Half precision. A raster is 0/1 planes and one ramp; fp16 is exact for
+        # the first and keeps about three decimals of the second, which an
+        # ablation could not distinguish from fp32 (2.8211 CE against 2.8441 for
+        # a 256-level quantization -- but that arm lost, so the margin is where
+        # fp16 sits, not below it). Halving the replay window's footprint is
+        # what lets it hold sixteen shards rather than five.
+        states=torch.from_numpy(states).half(),
         policies=torch.from_numpy(policies),
         policy_masks=torch.from_numpy(policy_masks).bool(),
         visits=torch.from_numpy(visits),
